+++ CHANGELOG.md
@@ -6,7 +6,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<<<<<<< HEAD
 ## [4.7.0](https://github.com/ASFHyP3/hyp3-gamma/compare/v4.6.0...v4.7.0)
 
 ### Added
@@ -16,13 +15,27 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   * `water_masking` parameter to `ifm_sentinel.insar_sentinel_gamma` function
   * `--apply-watger-mask` parameter to `insar` entrypoint
   * `-m` parameter to `ifm_sentinel.py` script
-=======
+
+## [4.6.3](https://github.com/ASFHyP3/hyp3-gamma/compare/v4.6.2...v4.6.3)
+
+### Added
+* create and output the incidence ellipsoidal angle geotiff file in the product
+
+## [4.6.2](https://github.com/ASFHyP3/hyp3-gamma/compare/v4.6.1...v4.6.2)
+
+### Added
+* InSAR products now include a land/water mask GeoTIFF and a corresponding ArcGIS metadata xml file
+
+### Changed
+* Upgraded to hyp3-metadata [v1.2.4](https://github.com/ASFHyP3/hyp3-metadata-templates/blob/develop/CHANGELOG.md#124)
+  from v1.2.3
 
 ## [4.6.1](https://github.com/ASFHyP3/hyp3-gamma/compare/v4.6.0...v4.6.1)
 
 ### Changed
 * `conda-env.yml` has been renamed to `environment.yml` in-line with community practice
->>>>>>> 8d9a3a28b6f597854ca8a72e0789fec4c6686846
+
+* Upgraded to hyp3-lib [v1.6.8](https://github.com/ASFHyP3/hyp3-lib/blob/develop/CHANGELOG.md#168) from v1.6.7
 
 ## [4.6.0](https://github.com/ASFHyP3/hyp3-gamma/compare/v4.5.1...v4.6.0)